+++ CoffiiMe/src/main/java/app/coffeeTypes/latteMacchiato/LatteMacchiato.java
@@ -8,9 +8,9 @@ public class LatteMacchiato implements Coffee {
 
     Color color;
     int volume;
-    Map<String, Integer> ingredients;
+    Map<String, Double> ingredients;
 
-    public LatteMacchiato(Color color, int volume, Map<String, Integer> ingredients) {
+    public LatteMacchiato(Color color, int volume, Map<String, Double> ingredients) {
         this.color = color;
         this.volume = volume;
         this.ingredients = ingredients;
@@ -27,7 +27,7 @@ public class LatteMacchiato implements Coffee {
     }
 
     @Override
-    public Map<String, Integer> getIngredients() {
+    public Map<String, Double> getIngredients() {
         return ingredients;
     }
 }
